Fixed build script

+++ build.sh
@@ -3,18 +3,25 @@
 start=false
 while getopts ":s" option; do
   case $option in
-    b) start=true ;;
+    s) start=true ;;
     ?) echo "error: option -$OPTARG is not implemented"; exit ;;
   esac
 done
+echo $start
 
-./compile rte/rte.java ThanOS/src/**/*.java -o boot
+./compile rte/rte/*.java rte/java/**/*.java ThanOS/src/**/*.java -o boot
+
+if [ $? -eq 1 ]
+  then
+    exit
+fi
+
 rm -R build/*
 mv BOOT_FLP.IMG build/BOOT_FLP.IMG
 mv syminfo.txt build/syminfo.txt
 
-if [ start ]
-  then
-    echo "Starting"
-    qemu-system-i386 -boot a -fda build/BOOT_FLP.IMG
-fi
+
+case $start in
+  (true)    qemu-system-i386 -boot a -fda build/BOOT_FLP.IMG;;
+  (false)   ;;
+esac
